refactor: share Buzz authenticated relay setup
